Add files via upload
An electronic health record (EHR) is a technology that allows you to keep track of your health information. It keeps computerized records of several healthcare organizations. Records are exchanged via enterprise-wide data systems as well as other networking technologies and exchanges. Patients nowadays expect immediate access to their health information. However, the health sector comes with immediate access to data, and there are worries about the privacy and security of medical records of patients. As a result, a blockchain-based solution can assist in resolving this issue. The blockchain has the potential to beat the conventional centralized system, which suffers from a severe lack of accessibility. This is a decentralized technology that has recently been presented to provide a new viewpoint on data security and system efficiency.
Paper Published - https://www.hindawi.com/journals/js/2022/7299185/